# openapi/updater-api.v1.yaml
...
openapi: 3.1.0
info:
  title: cicd-updater sidecar API
  version: "1.0"
  description: "HTTP API of the cicd-updater sidecar. Everything under /v1 requires `Authorization: Bearer <token>`. Errors are RFC 9457 problem documents (application/problem+json) with `type: urn:cicd-updater:problem:<code>`; codes: `unauthorized` (401), `not_found` (404), `unsupported_media_type` (415), `payload_too_large` (413), `invalid_request` (422), `release_not_found` (404), `release_unverifiable` (422), `release_refused` (409), `release_mismatch` (409), `source_not_allowed` (409), `busy` (409), `blocked` (409), `not_scheduled` (409), `point_of_no_return` (409), `not_finished` (409), `feed_unavailable` (502), `internal` (500). Clients must ignore unknown fields and render unknown codes generically."
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: http://updater:8090
security:
  - bearerToken: []
paths:
...
  /public/v1/status:
    get:
      operationId: getPublicStatus
      summary: What an anonymous visitor may see (for maintenance pages).
      security: []
      responses:
        "200":
          description: The public status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicStatus"
        "404":
          description: The public status is disabled.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /v1/state:
    get:
      operationId: getState
      summary: "Everything a UI needs: phase, run, history, running version, capabilities."
      parameters:
        - name: refresh
          in: query
          required: false
          schema:
            type: boolean
          description: Recompute instead of using the cache.
      responses:
        "200":
          description: The state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StateView"
        "401":
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/Problem"
  /v1/capabilities:
    get:
      operationId: getCapabilities
      summary: The preflight result only.
      parameters:
        - name: refresh
          in: query
          required: false
          schema:
            type: boolean
          description: Recompute instead of using the cache.
      responses:
        "200":
          description: The capabilities.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Capabilities"
        "401":
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/Problem"
  /v1/releases:
    get:
      operationId: listReleases
      summary: Newer releases from the configured feed, with refusals (metadata not verified).
      parameters:
        - name: refresh
          in: query
          required: false
          schema:
            type: boolean
          description: Recompute instead of using the cache.
      responses:
        "200":
          description: The releases.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReleasesView"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "502":
          description: The release host failed (feed_unavailable).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /v1/releases/{version}/verification:
    post:
      operationId: verifyRelease
      summary: "Dry run: fetch and verify release.json and the image signatures, no pull."
      parameters:
        - name: version
          in: path
          required: true
          schema:
            type: string
            maxLength: 64
      responses:
        "200":
          description: The verification result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: No such release or no release.json (release_not_found).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "422":
          description: The release does not verify (release_unverifiable).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "502":
          description: The release host failed (feed_unavailable).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /v1/runs:
    get:
      operationId: listRuns
      summary: History of runs (summaries, newest first).
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: The runs.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunsView"
        "401":
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/Problem"
    post:
      operationId: scheduleRun
      summary: Schedule a run; the release is fully verified before it is accepted.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleRequest"
      responses:
        "202":
          description: The run is scheduled (or running when it starts at once).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StateView"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: release_not_found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "409":
          description: busy, blocked, release_refused, release_mismatch or source_not_allowed
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "422":
          description: invalid_request or release_unverifiable
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "502":
          description: feed_unavailable
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /v1/runs/{runId}:
    get:
      operationId: getRun
      summary: One run including its log.
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: string
            pattern: ^r-\d{1,15}-[0-9a-f]{4}$
      responses:
        "200":
          description: The run.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: not_found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        default:
          $ref: "#/components/responses/Problem"
    patch:
      operationId: rescheduleRun
      summary: Move a scheduled run.
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: string
            pattern: ^r-\d{1,15}-[0-9a-f]{4}$
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RescheduleRequest"
      responses:
        "200":
          description: The run was moved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StateView"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: not_found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "409":
          description: not_scheduled
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "422":
          description: invalid_request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /v1/runs/{runId}/cancel:
    post:
      operationId: cancelRun
      summary: Cancel a scheduled run, or abort a running one before the point of no return.
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: string
            pattern: ^r-\d{1,15}-[0-9a-f]{4}$
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunActionRequest"
      responses:
        "200":
          description: Cancelled.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StateView"
        "202":
          description: Abort requested; the run stops at its next check point.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StateView"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: not_found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "409":
          description: point_of_no_return or not_scheduled
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /v1/runs/{runId}/acknowledge:
    post:
      operationId: acknowledgeRun
      summary: Clear a finished run (it stays in the history).
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: string
            pattern: ^r-\d{1,15}-[0-9a-f]{4}$
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunActionRequest"
      responses:
        "200":
          description: Acknowledged.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StateView"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: not_found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "409":
          description: not_finished
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /v1/events:
    get:
      operationId: listEvents
      summary: Journal events after a cursor, for the app's audit log (exactly once).
      parameters:
        - name: after
          in: query
          required: false
          schema:
            type: string
            pattern: ^\d{15}-\d{6}$
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
      responses:
        "200":
          description: The events.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventsView"
        "401":
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/Problem"
  /v1/backups:
    get:
      operationId: listBackups
      summary: Backups with metadata and protection flag (never their content).
      responses:
        "200":
          description: The backups.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BackupsView"
        "401":
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/Problem"
  /v1/config:
    get:
      operationId: getConfig
      summary: The effective configuration (redacted) and its hash.
      responses:
        "200":
          description: The configuration.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigView"
        "401":
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/Problem"
  /v1/openapi.json:
    get:
      operationId: getOpenApi
      summary: This document.
      responses:
        "200":
          description: The OpenAPI document.
          content:
            application/json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/Problem"
components:
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
  responses:
    Problem:
      description: An RFC 9457 problem document.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Unauthorized:
      description: Missing or wrong bearer token (unauthorized).
      headers:
        WWW-Authenticate:
          schema:
            type: string
          description: Bearer
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    Problem:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        code:
          type: string
        blockers:
          type: array
          items:
            $ref: "#/components/schemas/Blocker"
        errors:
          maxItems: 10
          type: array
          items:
            type: object
            properties:
              path:
                type: string
              message:
                type: string
            required:
              - path
              - message
        reasons:
          type: array
          items:
            type: string
        checks:
          $ref: "#/components/schemas/VerificationResult"
        feedError:
          type: string
      required:
        - type
        - title
        - status
        - code
      additionalProperties: {}
    PublicStatus:
      type: object
      properties:
        phase:
          type: string
          enum:
            - idle
            - scheduled
            - running
            - succeeded
            - failed
        runId:
          type:
            - string
            - "null"
        outcome:
          anyOf:
            - type: string
              enum:
                - succeeded
                - unchanged
                - rolled_back
                - needs_attention
            - type: "null"
        startsAt:
          anyOf:
            - type: string
              format: date-time
              pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
            - type: "null"
        startedAt:
          anyOf:
            - type: string
              format: date-time
              pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
            - type: "null"
        finishedAt:
          anyOf:
            - type: string
              format: date-time
              pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
            - type: "null"
        step:
          anyOf:
            - type: string
              enum:
                - prepare
                - fetch
                - backup
                - stop
                - migrate
                - start
                - health
                - smoke
                - finish
            - type: "null"
        steps:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                enum:
                  - prepare
                  - fetch
                  - backup
                  - stop
                  - migrate
                  - start
                  - health
                  - smoke
                  - finish
              status:
                type: string
                enum:
                  - pending
                  - running
                  - done
                  - failed
                  - skipped
            required:
              - id
              - status
        progress:
          type: integer
          minimum: 0
          maximum: 100
        message:
          anyOf:
            - $ref: "#/components/schemas/Message"
            - type: "null"
        failureCode:
          type:
            - string
            - "null"
        targetVersion:
          type: string
        fromVersion:
          type:
            - string
            - "null"
        serverTime:
          type: string
          format: date-time
          pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
      required:
        - phase
        - runId
        - outcome
        - startsAt
        - startedAt
        - finishedAt
        - step
        - steps
        - progress
        - message
        - failureCode
        - serverTime
...
    RescheduleRequest:
      type: object
      properties:
        leadSeconds:
          type: integer
          minimum: 0
        startsAt:
          type: string
          format: date-time
          pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
        requestedBy:
          $ref: "#/components/schemas/RequestedBy"
      additionalProperties: false
...
    RunActionRequest:
      type: object
      properties:
        requestedBy:
          $ref: "#/components/schemas/RequestedBy"
      additionalProperties: false
...
    RunsView:
      type: object
      properties:
        runs:
          type: array
          items:
            $ref: "#/components/schemas/RunSummary"
      required:
        - runs
...